+++ db/migrate/20190925094306_add_currently_teaching_to_student_loans_eligibilities.rb
@@ -1,0 +1,5 @@
+class AddCurrentlyTeachingToStudentLoansEligibilities < ActiveRecord::Migration[5.2]
+  def change
+    add_column :student_loans_eligibilities, :currently_teaching, :boolean
+  end
+end