+++ spec/features/missing_verify_information_spec.rb
@@ -4,8 +4,9 @@ RSpec.feature "Missing information from GOV.UK Verify" do
   scenario "Claimant is asked a payroll gender question when Verify doesn’t provide their gender" do
     claim = start_claim
     choose_qts_year
+    choose_currently_teaching
     choose_school schools(:penistone_grammar_school)
-    choose_still_teaching
+    choose_where_teaching
     choose_subjects_taught
     click_on "Continue"
     fill_in_repayment_amount "1100"
@@ -53,8 +54,9 @@ RSpec.feature "Missing information from GOV.UK Verify" do
   scenario "Claimant is asked an address question when Verify doesn’t provide their address" do
     claim = start_claim
     choose_qts_year
+    choose_currently_teaching
     choose_school schools(:penistone_grammar_school)
-    choose_still_teaching
+    choose_where_teaching
     choose_subjects_taught
     click_on "Continue"
     fill_in_repayment_amount "1100"