+++ spec/requests/claims_spec.rb
@@ -77,12 +77,11 @@ RSpec.describe "Claims", type: :request do
       before { post claims_path }
 
       it "renders a static ineligibility page" do
-        Claim.order(:created_at).last.eligibility.update(employment_status: "no_school")
+        Claim.order(:created_at).last.eligibility.update(currently_teaching: false)
 
         get claim_path("ineligible")
 
         expect(response.body).to include("You’re not eligible")
-        expect(response.body).to include("You can only get this payment if you’re still employed at a school.")
       end
     end