+++ spec/requests/timeout_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe "Claim session timing out", type: :request do
     end
 
     let(:current_claim) { Claim.order(:created_at).last }
-    let(:after_expiry) { timeout_length_in_minutes.minutes + 1.second }
+    let(:after_expiry) { timeout_length_in_minutes.minutes + 5.seconds }
 
     it "clears the session and redirects to the timeout page" do
       expect(session[:claim_id]).to eql current_claim.to_param
@@ -33,13 +33,13 @@ RSpec.describe "Claim session timing out", type: :request do
     end
 
     let(:current_claim) { Claim.order(:created_at).last }
-    let(:before_expiry) { timeout_length_in_minutes.minutes - 2.seconds }
+    let(:before_expiry) { timeout_length_in_minutes.minutes - 5.seconds }
 
     it "does not timeout the session" do
       travel before_expiry do
         put claim_path("qts-year"), params: {claim: {eligibility_attributes: {qts_award_year: "2014_2015"}}}
 
-        expect(response).to redirect_to(claim_path("claim-school"))
+        expect(response).to redirect_to(claim_path(PageSequence::SLUGS[1]))
         expect(session[:verify_request_id]).not_to be_nil
       end
     end